update benchmark recipe NEW

+++ SLIDE-x-BENCH/KERNEL/qsort/frst.c
@@ -43,7 +43,7 @@
 #include <stdint.h>
 #include <values.h>
 
-typedef double TARGET_TYPE;
+typedef int64_t TARGET_TYPE;
 typedef uint8_t TARGET_INDEX;
 
 #define SWAP(a,b) temp=(a);(a)=(b);(b)=temp;
